requirements.txt
openai~=1.59.4
python-dotenv~=1.0.1